Use the `response_url` to post messages back to Slack
Slack webhooks no longer allow you to override the default channel
for the webhook, which means we can't use a pre-registered hook to
reply to a user in the same place where they used the slash command.

Instead, we can use the `response_url` parameter from the request
that is received from Slack; this parameter contains the URL for a
temporary webhook that posts back to the same channel where the user
ran their slash command.